pyproject.toml
[tool.poetry]
name = "inso-extpipes-cli"
version = "2.2.1"
description = "A CLI to deploy CDF Extraction Pipelines"
authors = ["Peter Arwanitis <peter.arwanitis@cognite.com>", "Tugce Ozgur Oztetik <tugce.oztetik@cognite.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/cognitedata/inso-extpipes-cli"

packages = [
    { include="src" },
]

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[tool.black]
line-length = 120
target_version = ['py310']
include = '\.py$'

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility.html
profile = "black"
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["click", "pytest"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
pydantic = "^1.10.4"
dependency-injector = {version = "^4.41.0", extras = ["yaml"]}
click = "^8.1.3"
cognite-sdk = {version = "^5.2.0", extras = ["pandas"]}
setuptools = "^62.1.0"
rich = "^13.2.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
isort = "^5.11.4"
flake8 = "^6.0.0"
mypy = "^0.991"
# autoflake = "=1.3"
pre-commit = "^3.0.0"
pytest = "^7.2.1"
# pytest-cov = "^3.0.0"

[tool.poetry.scripts]
extpipes-cli = "src.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.semantic_release]
version_toml = [
"pyproject.toml:tool.poetry.version"
]
version_variable = [
    "src/__init__.py:__version__",
]
